add the trouble_deal task and change the firmwalk to support the all funsion.
the trouble_deal task have nothing code
the funsion which count the PWM num has split to the walk_contral task and name's StatDistance_task but haven't test.
the deal_ble_data funsion have test but not test crc_check

+++ USER/APP/walk-contral.c
@@ -29,7 +29,7 @@ void WalK_Init(void)
 //功能:小车行走动作
 //输入:行走方向
 //输出:
-void CR_Walk_Contral(u8 cmd)
+void CR_Walk_Contral(u32 cmd)
 {
     switch(cmd)
     {
